DFE-666 Move content db classes into separate project.
