+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api/Controllers/ContentController.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Linq;
-using GovUk.Education.ExploreEducationStatistics.Content.Api.Data;
 using GovUk.Education.ExploreEducationStatistics.Content.Api.Services;
 using GovUk.Education.ExploreEducationStatistics.Content.Api.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
