+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api/GovUk.Education.ExploreEducationStatistics.Content.Api.csproj
@@ -7,7 +7,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <Folder Include="Models" />
         <Folder Include="ViewModels" />
         <Folder Include="wwwroot\" />
     </ItemGroup>
