+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Converters/ContentBlockChartConverter.cs
@@ -1,35 +1,29 @@
 using System;
-using GovUk.Education.ExploreEducationStatistics.Content.Api.Models;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Converters
+namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Converters
 {
-    public class ContentBlockConverter : JsonConverter
+    public class ContentBlockChartConverter : JsonConverter
     {
+        
         public override bool CanWrite => false;
         public override bool CanRead => true;
         
         public override bool CanConvert(Type objectType)
         {
-            return (objectType == typeof(IContentBlock));
+            return (objectType == typeof(IContentBlockChart));
         }
-        
+
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
             var jsonObject = JObject.Load(reader);
-            var contentBlock = default(IContentBlock);
+            var contentBlock = default(IContentBlockChart);
             
             switch (jsonObject["Type"].Value<string>())
             {
-                case "MarkDownBlock":
-                    contentBlock = new MarkDownBlock();
-                    break;
-                case "InsetTextBlock":
-                    contentBlock = new InsetTextBlock();
-                    break;
-                case "DataBlock":
-                    contentBlock = new DataBlock();
+                case "line":
+                    contentBlock = new LineChart();
                     break;
             }
             serializer.Populate(jsonObject.CreateReader(), contentBlock);
@@ -41,5 +35,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Converters
             throw new InvalidOperationException("Use default serialization.");
             // serializer.Serialize(writer, value, typeof(MarkDownBlock));
         }
+
+        
+        
     }
 }
