+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Link.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Models
+namespace GovUk.Education.ExploreEducationStatistics.Content.Model
 {
     public class Link
     {
