+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Publication.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Models
+namespace GovUk.Education.ExploreEducationStatistics.Content.Model
 {
     public class Publication
     {
