+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Update.cs
@@ -1,7 +1,7 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Models
+namespace GovUk.Education.ExploreEducationStatistics.Content.Model
 {
     public class Update
     {
